+++ app/flag_provider.go
@@ -9,7 +9,6 @@ import (
 	"github.com/adrianliechti/devkube/provider/none"
 
 	dockercli "github.com/adrianliechti/devkube/pkg/docker"
-	kindcli "github.com/adrianliechti/devkube/pkg/kind"
 )
 
 var ProviderFlag = &cli.StringFlag{
@@ -25,10 +24,6 @@ func Provider(c *cli.Context) (provider.Provider, error) {
 		return none.New(), nil
 	case "kind", "":
 		if _, _, err := dockercli.Info(c.Context); err != nil {
-			return nil, err
-		}
-
-		if _, _, err := kindcli.Info(c.Context); err != nil {
 			return nil, err
 		}
